feat: ping-3172 fix pr
