+++ controllers/chat/index.js
@@ -1,11 +1,13 @@
 const deleteChat = require('./deleteChat');
 const sendSignedMessage = require('./sendSignedMessage');
+const moveToAnon = require('./moveToAnon');
 const setSignedChannelAsRead = require('./setSignedChannelAsRead');
 const findOrCreateChannelBySignedSender = require('./findOrCreateChannelBySignedSender');
 
 module.exports = {
   deleteChat,
   sendSignedMessage,
+  moveToAnon,
   setSignedChannelAsRead,
   findOrCreateChannelBySignedSender
 };
